Export au format excel du résultat de la recherche de communes, contrôleur commune séparé
